requirements.txt
alabaster
argon2-cffi
argon2-cffi-bindings
asgiref
async-generator
attrs
Babel
certifi
cffi
charset-normalizer
colorama
coverage
cryptography
defusedxml
distlib
Django
django-allauth
django-cors-headers
django-debug-toolbar
django-filter
django-guardian
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djangorestframework-xml
djangorestframework-yaml
djoser
docutils
exceptiongroup
filelock
h11
h2
hpack
httpie
hypercorn
hyperframe
idna
imagesize
Jinja2
Markdown
markdown-it-py
MarkupSafe
mdurl
multidict
numpy
oauthlib
outcome
packaging
pandas
Pillow
pipenv
platformdirs
priority
pycparser
Pygments
PyJWT
PySocks
python-dateutil
python-dotenv
python3-openid
pytz
PyYAML
reportlab
requests
requests-oauthlib
requests-toolbelt
rest-framework-simplejwt
rich
selenium
six
snakeviz
sniffio
snowballstemmer
social-auth-app-django
social-auth-core
sortedcontainers
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
sqlparse
toml
tornado
trio
trio-websocket
tweepy
tzdata
urllib3
virtualenv
virtualenv-clone
wsproto